Move a bunch of helpers
AddIntFlag AddDurationFlag AddStringFlag checkExperimental createBuildContext testComposeUp requiresStargz newJWEKeyPair rmiAll
