+++ cmd/nerdctl/helpers/testing.go
@@ -17,8 +17,21 @@
 package helpers
 
 import (
+	"context"
 	"net"
+	"os"
+	"os/exec"
+	"path/filepath"
 	"strings"
+	"testing"
+
+	"gotest.tools/v3/assert"
+
+	containerd "github.com/containerd/containerd/v2/client"
+	"github.com/containerd/containerd/v2/core/content"
+
+	"github.com/containerd/nerdctl/v2/pkg/buildkitutil"
+	"github.com/containerd/nerdctl/v2/pkg/testutil"
 )
 
 func FindIPv6(output string) net.IP {
@@ -34,4 +47,95 @@ func FindIPv6(output string) net.IP {
 		}
 	}
 	return net.ParseIP(ipv6)
+}
+
+func RequiresStargz(base *testutil.Base) {
+	info := base.Info()
+	for _, p := range info.Plugins.Storage {
+		if p == "stargz" {
+			return
+		}
+	}
+	base.T.Skip("test requires stargz")
+}
+
+type JweKeyPair struct {
+	Prv     string
+	Pub     string
+	Cleanup func()
+}
+
+func NewJWEKeyPair(t testing.TB) *JweKeyPair {
+	testutil.RequireExecutable(t, "openssl")
+	td, err := os.MkdirTemp(t.TempDir(), "jwe-key-pair")
+	assert.NilError(t, err)
+	prv := filepath.Join(td, "mykey.pem")
+	pub := filepath.Join(td, "mypubkey.pem")
+	cmds := [][]string{
+		// Exec openssl commands to ensure that nerdctl is compatible with the output of openssl commands.
+		// Do NOT refactor this function to use "crypto/rsa" stdlib.
+		{"openssl", "genrsa", "-out", prv},
+		{"openssl", "rsa", "-in", prv, "-pubout", "-out", pub},
+	}
+	for _, f := range cmds {
+		cmd := exec.Command(f[0], f[1:]...)
+		if out, err := cmd.CombinedOutput(); err != nil {
+			t.Fatalf("failed to run %v: %v (%q)", cmd.Args, err, string(out))
+		}
+	}
+	return &JweKeyPair{
+		Prv: prv,
+		Pub: pub,
+		Cleanup: func() {
+			_ = os.RemoveAll(td)
+		},
+	}
+}
+
+func RmiAll(base *testutil.Base) {
+	base.T.Logf("Pruning images")
+	imageIDs := base.Cmd("images", "--no-trunc", "-a", "-q").OutLines()
+	// remove empty output line at the end
+	imageIDs = imageIDs[:len(imageIDs)-1]
+	// use `Run` on purpose (same below) because `rmi all` may fail on individual
+	// image id that has an expected running container (e.g. a registry)
+	base.Cmd(append([]string{"rmi", "-f"}, imageIDs...)...).Run()
+
+	base.T.Logf("Pruning build caches")
+	if _, err := buildkitutil.GetBuildkitHost(testutil.Namespace); err == nil {
+		base.Cmd("builder", "prune", "--force").AssertOK()
+	}
+
+	// For BuildKit >= 0.11, pruning cache isn't enough to remove manifest blobs that are referred by build history blobs
+	// https://github.com/containerd/nerdctl/pull/1833
+	if base.Target == testutil.Nerdctl {
+		base.T.Logf("Pruning all content blobs")
+		addr := base.ContainerdAddress()
+		client, err := containerd.New(addr, containerd.WithDefaultNamespace(testutil.Namespace))
+		assert.NilError(base.T, err)
+		cs := client.ContentStore()
+		ctx := context.TODO()
+		wf := func(info content.Info) error {
+			base.T.Logf("Pruning blob %+v", info)
+			if err := cs.Delete(ctx, info.Digest); err != nil {
+				base.T.Log(err)
+			}
+			return nil
+		}
+		if err := cs.Walk(ctx, wf); err != nil {
+			base.T.Log(err)
+		}
+
+		base.T.Logf("Pruning all images (again?)")
+		imageIDs = base.Cmd("images", "--no-trunc", "-a", "-q").OutLines()
+		base.T.Logf("pruning following images: %+v", imageIDs)
+		base.Cmd(append([]string{"rmi", "-f"}, imageIDs...)...).Run()
+	}
+}
+
+func CreateBuildContext(t *testing.T, dockerfile string) string {
+	tmpDir := t.TempDir()
+	err := os.WriteFile(filepath.Join(tmpDir, "Dockerfile"), []byte(dockerfile), 0644)
+	assert.NilError(t, err)
+	return tmpDir
 }
