+++ cmd/nerdctl/helpers/testing_linux.go
@@ -1,0 +1,79 @@
+/*
+   Copyright The containerd Authors.
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+*/
+
+package helpers
+
+import (
+	"fmt"
+	"io"
+	"strings"
+	"testing"
+	"time"
+
+	"github.com/containerd/nerdctl/v2/pkg/testutil"
+	"github.com/containerd/nerdctl/v2/pkg/testutil/nettestutil"
+)
+
+func ComposeUp(t *testing.T, base *testutil.Base, dockerComposeYAML string, opts ...string) {
+	comp := testutil.NewComposeDir(t, dockerComposeYAML)
+	defer comp.CleanUp()
+
+	projectName := comp.ProjectName()
+	t.Logf("projectName=%q", projectName)
+
+	base.ComposeCmd(append(append([]string{"-f", comp.YAMLFullPath()}, opts...), "up", "-d")...).AssertOK()
+	defer base.ComposeCmd("-f", comp.YAMLFullPath(), "down", "-v").Run()
+	base.Cmd("volume", "inspect", fmt.Sprintf("%s_db", projectName)).AssertOK()
+	base.Cmd("network", "inspect", fmt.Sprintf("%s_default", projectName)).AssertOK()
+
+	checkWordpress := func() error {
+		resp, err := nettestutil.HTTPGet("http://127.0.0.1:8080", 10, false)
+		if err != nil {
+			return err
+		}
+		respBody, err := io.ReadAll(resp.Body)
+		if err != nil {
+			return err
+		}
+		if !strings.Contains(string(respBody), testutil.WordpressIndexHTMLSnippet) {
+			t.Logf("respBody=%q", respBody)
+			return fmt.Errorf("respBody does not contain %q", testutil.WordpressIndexHTMLSnippet)
+		}
+		return nil
+	}
+
+	var wordpressWorking bool
+	for i := 0; i < 30; i++ {
+		t.Logf("(retry %d)", i)
+		err := checkWordpress()
+		if err == nil {
+			wordpressWorking = true
+			break
+		}
+		// NOTE: "<h1>Error establishing a database connection</h1>" is expected for the first few iterations
+		t.Log(err)
+		time.Sleep(3 * time.Second)
+	}
+
+	if !wordpressWorking {
+		t.Fatal("wordpress is not working")
+	}
+	t.Log("wordpress seems functional")
+
+	base.ComposeCmd("-f", comp.YAMLFullPath(), "down", "-v").AssertOK()
+	base.Cmd("volume", "inspect", fmt.Sprintf("%s_db", projectName)).AssertFail()
+	base.Cmd("network", "inspect", fmt.Sprintf("%s_default", projectName)).AssertFail()
+}
